+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/Era.hs
@@ -27,7 +27,7 @@ module Cardano.Ledger.Shelley.Era (
 where
 
 import Cardano.Ledger.Coin (Coin)
-import Cardano.Ledger.Core (Era (..), EraRule, TranslationContext, Value)
+import Cardano.Ledger.Core (Era (..), EraRule, Value)
 import Cardano.Ledger.Crypto as CC (Crypto)
 
 data ShelleyEra c
@@ -37,8 +37,6 @@ instance CC.Crypto c => Era (ShelleyEra c) where
   type ProtVerLow (ShelleyEra c) = 2
 
 type instance Value (ShelleyEra _c) = Coin
-
-type instance TranslationContext (ShelleyEra c) = ()
 
 data ShelleyBBODY era