+++ libs/cardano-protocol-tpraos/src/Cardano/Protocol/TPraos/API.hs
@@ -75,7 +75,6 @@ import Cardano.Ledger.Keys (
 import Cardano.Ledger.Mary (MaryEra)
 import Cardano.Ledger.PoolDistr (PoolDistr (..), individualPoolStake)
 import Cardano.Ledger.Shelley (ShelleyEra)
-import Cardano.Ledger.Shelley.Genesis (ShelleyGenesis (..))
 import Cardano.Ledger.Shelley.LedgerState (
   EpochState (..),
   NewEpochState (..),
@@ -85,6 +84,7 @@ import Cardano.Ledger.Shelley.LedgerState (
  )
 import Cardano.Ledger.Shelley.PParams (ShelleyPParamsHKD (..))
 import Cardano.Ledger.Shelley.Rules (ShelleyTickfPredFailure)
+import Cardano.Ledger.Shelley.Translation (FromByronTranslationContext (..))
 import Cardano.Ledger.Slot (SlotNo)
 import Cardano.Protocol.TPraos.BHeader (
   BHBody,
@@ -598,14 +598,14 @@ getLeaderSchedule globals ss cds poolHash key pp = Set.filter isLeader epochSlot
 -- way as 'translateToShelleyLedgerState'.
 mkInitialShelleyLedgerView ::
   forall c.
-  ShelleyGenesis c ->
+  FromByronTranslationContext c ->
   LedgerView c
-mkInitialShelleyLedgerView genesisShelley =
-  let !ee = _extraEntropy . sgProtocolParams $ genesisShelley
+mkInitialShelleyLedgerView transCtxt =
+  let !ee = _extraEntropy . fbtcProtocolParams $ transCtxt
    in LedgerView
-        { lvD = _d . sgProtocolParams $ genesisShelley
+        { lvD = _d . fbtcProtocolParams $ transCtxt
         , lvExtraEntropy = ee
         , lvPoolDistr = PoolDistr Map.empty
-        , lvGenDelegs = GenDelegs $ sgGenDelegs genesisShelley
-        , lvChainChecks = pparamsToChainChecksPParams . sgProtocolParams $ genesisShelley
+        , lvGenDelegs = GenDelegs $ fbtcGenDelegs transCtxt
+        , lvChainChecks = pparamsToChainChecksPParams . fbtcProtocolParams $ transCtxt
         }